requirements.txt
Appium-Python-Client~=1.0.2
selenium~=3.141.0
pytest==7.1.2
allure-pytest
pytest-parallel
pytest-xdist
pytest-rerunfailures
pytest-sugar
pytest-json
pymysql~=0.10.1
sshtunnel~=0.1.5
requests~=2.24.0
invoke~=1.4.1
python-dotenv
pytest-html-reporter~=0.2.3
webdriver-manager
allure-python-commons
pytest-ordering